decompose topics updater
join topic parsing functions together

refact

fix comment number
