--- tests/test_identify_updates_of_topics/test_folder_updater.py
@@ -1,129 +1,0 @@
-from copy import deepcopy
-from pathlib import Path
-
-import pytest
-
-from _dependencies.commons import ChangeType
-from identify_updates_of_topics._utils.folder_updater import SearchUpdater
-from identify_updates_of_topics._utils.forum import ForumClient
-from tests.common import find_model
-from tests.factories import db_factories, db_models
-from tests.test_identify_updates_of_topics.factories import SearchSummaryFactory, fake
-
-
-class PatchedSearchUpdater(SearchUpdater):
-    forum: 'FakeForum'
-
-
-@pytest.fixture()
-def searches_updater(db_client) -> PatchedSearchUpdater:
-    return PatchedSearchUpdater(db_client, FakeForum())
-
-
-class FakeForum(ForumClient):
-    def _get_comment_content(self, search_num: int, comment_num: int) -> bytes:
-        return Path('tests/fixtures/forum_comment.html').read_bytes()
-
-    def _get_topic_content(self, search_num: int) -> bytes:
-        return Path('tests/fixtures/forum_topic.html').read_bytes()
-
-
-class TestFolderUpdaterChangeLogCreation:
-    def test_new_search(self, session, searches_updater: PatchedSearchUpdater):
-        search_summary = SearchSummaryFactory.build()
-
-        change_log_ids = searches_updater._update_change_log_and_search(search_summary)
-
-        assert len(change_log_ids) == 1
-        assert find_model(session, db_models.ChangeLog, id=change_log_ids[0], change_type=ChangeType.topic_new)
-        assert find_model(session, db_models.Search, search_forum_num=search_summary.topic_id)
-
-    def test_changed_search(self, session, searches_updater: PatchedSearchUpdater):
-        start_replies = 2
-        search_summary = SearchSummaryFactory.build(
-            num_of_replies=start_replies + 1,
-        )
-        existed_search = db_factories.SearchFactory.create_sync(
-            search_forum_num=search_summary.topic_id,
-            forum_folder_id=search_summary.folder_id,
-            num_of_replies=start_replies,
-        )
-
-        change_log_ids = searches_updater._update_change_log_and_search(search_summary)
-
-        assert len(change_log_ids) == 4
-        assert find_model(
-            session,
-            db_models.ChangeLog,
-            id=change_log_ids[0],
-            change_type=ChangeType.topic_status_change,
-        )
-        assert find_model(
-            session,
-            db_models.ChangeLog,
-            id=change_log_ids[1],
-            change_type=ChangeType.topic_title_change,
-        )
-        assert find_model(
-            session,
-            db_models.ChangeLog,
-            id=change_log_ids[2],
-            change_type=ChangeType.topic_comment_new,
-        )
-        assert find_model(
-            session,
-            db_models.ChangeLog,
-            id=change_log_ids[3],
-            change_type=ChangeType.topic_inforg_comment_new,
-        )
-
-
-class TestFolderUpdaterChangeDetection:
-    def test_no_changes(self, searches_updater: PatchedSearchUpdater):
-        snapshot = SearchSummaryFactory.build()
-        search = deepcopy(snapshot)
-
-        changes = searches_updater._detect_changes(snapshot, search, False)
-
-        assert not changes
-
-    def test_changed_title(self, searches_updater: PatchedSearchUpdater):
-        snapshot = SearchSummaryFactory.build()
-        search = deepcopy(snapshot)
-        search.title = 'New Title'
-
-        changes = searches_updater._detect_changes(snapshot, search, False)
-
-        assert len(changes) == 1
-        assert changes[0].change_type == ChangeType.topic_title_change
-
-    def test_changed_status(self, searches_updater: PatchedSearchUpdater):
-        snapshot = SearchSummaryFactory.build()
-        search = deepcopy(snapshot)
-        search.status = 'New Status'
-
-        changes = searches_updater._detect_changes(snapshot, search, False)
-
-        assert len(changes) == 1
-        assert changes[0].change_type == ChangeType.topic_status_change
-
-    def test_changed_num_of_replies_no_inforg(self, searches_updater: PatchedSearchUpdater):
-        snapshot = SearchSummaryFactory.build()
-        search = deepcopy(snapshot)
-        search.num_of_replies -= 1
-
-        changes = searches_updater._detect_changes(snapshot, search, False)
-
-        assert len(changes) == 1
-        assert changes[0].change_type == ChangeType.topic_comment_new
-
-    def test_changed_num_of_replies_inforg(self, searches_updater: PatchedSearchUpdater):
-        snapshot = SearchSummaryFactory.build()
-        search = deepcopy(snapshot)
-        search.num_of_replies -= 1
-
-        changes = searches_updater._detect_changes(snapshot, search, True)
-
-        assert len(changes) == 2
-        assert changes[0].change_type == ChangeType.topic_comment_new
-        assert changes[1].change_type == ChangeType.topic_inforg_comment_new
