+++ tests/tools/run_identify_updates_of_topics.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock, patch
 
 from _dependencies import pubsub
 from identify_updates_of_topics import main
-from identify_updates_of_topics._utils import folder_updater
+from identify_updates_of_topics._utils import topic_updater
 from tests.common import get_dotenv_config, get_event_with_data, setup_logging_to_console
 from title_recognize.main import recognize_title
 
@@ -21,7 +21,7 @@ if __name__ == '__main__':
     with (
         patch('_dependencies.commons._get_config', get_dotenv_config),
         patch.object(pubsub, 'publish_to_pubsub', fake_publish_to_pubsub),
-        patch.object(folder_updater, 'recognize_title_via_api', fake_recognize_title_via_api),
+        patch.object(topic_updater, 'recognize_title_via_api', fake_recognize_title_via_api),
     ):
         folders = [(276, None), (120, None)]
         data = get_event_with_data(folders)
